[shuffle] Rename example Move pkg main/
main/ acts as the entry point for compilation, deployment, and code
generation. Libraries (AnotherPackageA) are only handled if pull in by main.

Closes: #9272
